refactor: extract new training utilities
